packages/rag_shared/pyproject.toml
[project]
name = "rag-shared"
version = "0.1.0"
description = "Shared utilities for the RAG system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "celery>=5.3",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "weaviate-client>=3.26.7,<4.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
